DASHTreeTest: fix 'unsigned char' printing
Consider the following minimal reproducer:

================================

enum ContainerType : uint8_t
{
  CONTAINERTYPE_NOTYPE,
  CONTAINERTYPE_INVALID,
  CONTAINERTYPE_MP4,
  CONTAINERTYPE_TS,
  CONTAINERTYPE_ADTS,
  CONTAINERTYPE_WEBM,
  CONTAINERTYPE_MATROSKA,
  CONTAINERTYPE_TEXT,
};

enum
{
  ENCRYTIONSTATE_UNENCRYPTED = 0,
  ENCRYTIONSTATE_ENCRYPTED = 1,
  ENCRYTIONSTATE_SUPPORTED = 2
};

int main()
{
	ContainerType ctype = CONTAINERTYPE_TS;

	std::cerr << "ctype = " << ctype << std::endl;
	std::cerr << "static_cast<unsigned int>(ctype) = " << static_cast<unsigned int>(ctype) << std::endl;
	return 0;
}

================================

GCC 8.3 puts out:

================================

$ /tmp/test
ctype = 3
static_cast<unsigned int>(ctype) = 3

================================

While GCC 10.2 puts out:

================================

ctype = <unprintable char 0x3>
static_cast<unsigned int>(ctype) = 3

================================

Fix the behavior by casting 'unsigned char' / 'uint8_t' to 'unsigned int'
in the stringstream buffering code.

+++ src/parser/DASHTreeTest.cpp
@@ -119,14 +119,14 @@ int main(int argc, char** argv)
       for (const adaptive::AdaptiveTree::Representation* rep : adp->representations_)
       {
         sstreamCur << "\t\t\tRep: aspect_: " << rep->aspect_ << ", bandwidth_ : " << rep->bandwidth_
-                  << ", channelCount_ : " << rep->channelCount_ << ", codecs_ : " << rep->codecs_
+                  << ", channelCount_ : " << static_cast<unsigned int>(rep->channelCount_) << ", codecs_ : " << rep->codecs_
                   << ", codec_private_data_ : " << rep->codec_private_data_
-                  << ", containerType_ : " << rep->containerType_ << ", duration_ : " << rep->duration_
+                  << ", containerType_ : " << static_cast<unsigned int>(rep->containerType_) << ", duration_ : " << rep->duration_
                   << ", flags_ : " << rep->flags_ << ", fpsRate_ : " << rep->fpsRate_
                   << ", fpsScale_ : " << rep->fpsScale_ << ", hdcpVersion_ : " << rep->hdcpVersion_
                   << ", height_ : " << rep->height_ << ", id : " << rep->id
                   << ", indexRangeMax_ : " << rep->indexRangeMax_ << ", indexRangeMin_ : " << rep->indexRangeMin_
-                  << ", nalLengthSize_ : " << rep->nalLengthSize_ << ", pssh_set_ : " << rep->pssh_set_
+                  << ", nalLengthSize_ : " << static_cast<unsigned int>(rep->nalLengthSize_) << ", pssh_set_ : " << rep->pssh_set_
                   << ", ptsOffset_ : " << rep->ptsOffset_ << ", samplingRate_ : " << rep->samplingRate_
                   << ", source_url_ : " << rep->source_url_ << ", startNumber_ : " << rep->startNumber_
                   << ", timescale_ : " << rep->timescale_ << ", timescale_ext_ : " << rep->timescale_ext_
